requirements.txt
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
asn1crypto==1.4.0
attrs==20.3.0
bcrypt==3.2.0
Beaker==1.11.0
breezy==3.1.0
btrfsutil==5.11.1
CacheControl==0.12.6
cached-property==1.5.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
cryptography==3.4.7
distlib==0.3.1
distro==1.5.0
dnspython==1.16.0
docker==4.4.4
docker-compose==1.28.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
gufw==21.4.0
html5lib==1.1
idna==3.1
importlib-metadata==3.7.3
jsonschema==3.2.0
keyutils==0.6
lensfun==0.3.95
LibAppArmor==3.0.1
lightdm-gtk-greeter-settings==1.2.2
louis==3.17.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
menulibre==2.2.1
mercurial==5.7.1
msgpack==1.0.2
mugshot==0.4.3
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.20.1
openshot-qt==2.5.1
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.0
paramiko==2.7.2
patiencediff==0.2.1
pep517==0.10.0
pexpect==4.8.0
pipenv==2020.11.15
ply==3.11
progress==1.5
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.20.0
pycparser==2.20
PyGObject==3.40.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-distutils-extra==2.39
python-dotenv==0.16.0
python-libtorrent==1.2.11
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.0.3
rados==2.0.0
rbd==2.0.0
requests==2.7.0
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
scour==0.38.2
six==1.15.0
team==1.0
terminator==2.1.0
texttable==1.6.3
toml==0.10.2
udiskie==2.3.3
ufw==0.36
urllib3==1.26.4
validate==5.1.0.dev0
virtualenv==20.3.1
virtualenv-clone==0.5.4
webencodings==0.5.1
websocket-client==0.58.0
zipp==3.4.1
